# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Contrib-Cycle",
        "contact": {
            "email": "jadiael1@gmail.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/admin/projects/{project}/payment-methods": {
            "get": {
                "tags": [
                    "Admin Payment Methods"
                ],
                "summary": "List payment methods",
                "description": "Returns payment methods for a project.",
                "operationId": "c4fd97b89ea9ec3fb7678d74d1cebd8c",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "Project ID.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment method list.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "data"
                                    ],
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CollectiveProjectPaymentMethodResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Project not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Payment Methods"
                ],
                "summary": "Create payment method",
                "description": "Adds a payment method to the project.",
                "operationId": "efa9c6cdc6aaf28af96c32f62959fb6d",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "Project ID.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCollectiveProjectPaymentMethodRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Payment method created.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "data"
                                    ],
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CollectiveProjectPaymentMethodResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Project not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/projects/{project}/payment-methods/{paymentMethod}": {
            "delete": {
                "tags": [
                    "Admin Payment Methods"
                ],
                "summary": "Delete payment method",
                "description": "Permanently removes a payment method.",
                "operationId": "cd54909737e88a42ab9eb3402243f9b6",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "Project ID.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "paymentMethod",
                        "in": "path",
                        "description": "Payment method ID.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment method deleted.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payment method not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Admin Payment Methods"
                ],
                "summary": "Update payment method",
                "description": "Updates a payment method for the project.",
                "operationId": "b1ce741790cbe8a86a05193ef5da55d4",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "Project ID.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "paymentMethod",
                        "in": "path",
                        "description": "Payment method ID.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCollectiveProjectPaymentMethodRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment method updated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "data"
                                    ],
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CollectiveProjectPaymentMethodResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payment method not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/projects/{project}/payment-methods/{paymentMethod}/deactivate": {
            "post": {
                "tags": [
                    "Admin Payment Methods"
                ],
                "summary": "Deactivate payment method",
                "description": "Deactivates a payment method. The project must keep at least one active method.",
                "operationId": "7e90733f108e00838184d47cabb70d1f",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "Project ID.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "paymentMethod",
                        "in": "path",
                        "description": "Payment method ID.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment method deactivated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Cannot deactivate last active method.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payment method not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/projects/{project}/payment-methods/{paymentMethod}/restore": {
            "post": {
                "tags": [
                    "Admin Payment Methods"
                ],
                "summary": "Activate payment method",
                "description": "Activates a payment method.",
                "operationId": "375f543e292916bbc5ff275aad868def",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "Project ID.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "paymentMethod",
                        "in": "path",
                        "description": "Payment method ID.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment method activated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payment method not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CollectiveProjectPaymentMethodResource": {
                "required": [
                    "id",
                    "type",
                    "sort_order",
                    "is_active"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 12
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "pix",
                            "bank_transfer"
                        ],
                        "example": "pix"
                    },
                    "label": {
                        "type": "string",
                        "example": "Primary",
                        "nullable": true
                    },
                    "sort_order": {
                        "type": "integer",
                        "example": 1
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "payload": {
                        "description": "Returned only for authenticated tokens with access to the payload.",
                        "nullable": true,
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/PaymentMethodPayloadPix"
                            },
                            {
                                "$ref": "#/components/schemas/PaymentMethodPayloadBankTransfer"
                            }
                        ]
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Invalid credentials."
                    }
                },
                "type": "object"
            },
...
            "MessageResponse": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Logged out."
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use the Bearer token returned by the login endpoints.",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}